pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srt-translator"
version = "1.1.0"
description = "基于 LLM 的异步字幕翻译工具 / Async LLM-powered subtitle translator"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "wjin999"}
]
keywords = [
    "subtitle",
    "translation",
    "srt",
    "llm",
    "deepseek",
    "nlp",
    "spacy",
    "async"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "openai>=1.0.0",
    "tqdm>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
merge = [
    "spacy>=3.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]
all = [
    "srt-translator[merge,dev]",
]

[project.scripts]
srt-translator = "srt_translator.cli:main"

[project.urls]
Homepage = "https://github.com/wjin999/srt-translator"
Repository = "https://github.com/wjin999/srt-translator"
Issues = "https://github.com/wjin999/srt-translator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
